refactor: remove everruns-contracts crate
Migration of types from everruns-contracts to appropriate locations:

**Domain types → everruns-core:**
- Agent, AgentStatus
- Session, SessionStatus
- LlmProvider (entity), LlmProviderType, LlmProviderStatus
- LlmModel, LlmModelStatus, LlmModelWithProvider
- Event
- CapabilityInfo (renamed from Capability to avoid conflict with trait)
- AgentCapability

**API-only types → everruns-api:**
- ListResponse<T>
- UpdateAgentCapabilitiesRequest
- CreateEventRequest

**Key changes:**
- Core types now support OpenAPI via `#[cfg_attr(feature = "openapi", derive(ToSchema))]`
- LlmProvider entity is in llm_entities module (not re-exported at root to avoid conflict with LlmProvider trait)
- Capability DTO renamed to CapabilityInfo to avoid conflict with Capability trait
- Removed contracts dependency from all crates

This consolidates the crate structure and eliminates the intermediate contracts layer.
