+++ crates/everruns-api/src/messages.rs
@@ -11,8 +11,9 @@ use axum::{
     Json, Router,
 };
 use chrono::{DateTime, Utc};
-use everruns_contracts::ListResponse;
 use everruns_storage::Database;
+
+use crate::common::ListResponse;
 use everruns_worker::AgentRunner;
 use serde::{Deserialize, Serialize};
 use std::{collections::HashMap, sync::Arc};
