+++ crates/everruns-api/src/users.rs
@@ -8,8 +8,9 @@ use axum::{
     Json, Router,
 };
 use chrono::{DateTime, Utc};
-use everruns_contracts::ListResponse;
 use everruns_storage::Database;
+
+use crate::common::ListResponse;
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
 use utoipa::ToSchema;
