--- crates/everruns-contracts/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "everruns-contracts"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[dependencies]
-serde.workspace = true
-serde_json.workspace = true
-uuid.workspace = true
-chrono.workspace = true
-utoipa.workspace = true
-
-# Re-export runtime types from core for backward compatibility
-everruns-core = { path = "../everruns-core" }
