--- crates/everruns-contracts/src/lib.rs
@@ -1,21 +1,0 @@
-// Public contracts for Everruns API
-// This crate defines DTOs, AG-UI event types, and JSON schemas
-// M2: Agent/Session/Messages model with Events as SSE notification channel
-//
-// Note: Message API contracts have been moved to everruns-api crate
-
-pub mod agent;
-pub mod capability;
-pub mod common;
-pub mod events;
-pub mod llm;
-pub mod session;
-pub mod tools;
-
-pub use agent::*;
-pub use capability::*;
-pub use common::*;
-pub use events::*;
-pub use llm::*;
-pub use session::*;
-pub use tools::*;
