+++ crates/everruns-core/src/event.rs
@@ -1,0 +1,23 @@
+// Event entity type
+//
+// This type represents an SSE notification record stored in the database.
+// Note: This is separate from events.rs which defines LoopEvent.
+
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+
+#[cfg(feature = "openapi")]
+use utoipa::ToSchema;
+
+/// Event - SSE notification record stored in database
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "openapi", derive(ToSchema))]
+pub struct Event {
+    pub id: Uuid,
+    pub session_id: Uuid,
+    pub sequence: i32,
+    pub event_type: String,
+    pub data: serde_json::Value,
+    pub created_at: DateTime<Utc>,
+}
