+++ crates/everruns-storage/Cargo.toml
@@ -22,5 +22,4 @@ sha2.workspace = true
 hex.workspace = true
 async-trait.workspace = true
 
-everruns-contracts = { path = "../everruns-contracts" }
 everruns-core = { path = "../everruns-core" }
